requirements.txt
python-dotenv
pyrogram
tgcrypto
aiofiles